Stock prediction of tcs
For the stock prediction for the internship in Bharat intern i would be taking the data of Tcs from kaggle to make the stock prediction analysis.
I also used LSTM model for the prediction Tata Consultancy Services (TCS) is one of the world's leading IT services and consulting companies. To assess the organization's current strategic position and make informed decisions for its future, an analysis using Internal and External Strategic Management Tools (ISMT) is imperative. This report provides a comprehensive analysis of TCS, taking into account internal factors such as strengths and weaknesses and external factors including opportunities and threats.